requirements.txt
flask
flask-cors
openai-whisper
google-generativeai
gTTS
